time laps is saved on indexedDB
On every change of time type (from pomodoro to break and vice versa) if time is bigger than 0 seconds, an object is saved to indexedDB with a unique id, type and time. set seconds was moved to useApp, so seconds would be accessible on save. A wrapper made for time type change, so there would be access to the old state